Add auth, search an artist and album page

+++ src/main/java/com/github/enteraname74/musik/controller/PlaylistController.java
@@ -1,28 +1,40 @@
 package com.github.enteraname74.musik.controller;
 
+import com.github.enteraname74.musik.controller.utils.ControllerUtils;
 import com.github.enteraname74.musik.domain.model.Playlist;
+import com.github.enteraname74.musik.domain.service.AuthService;
 import com.github.enteraname74.musik.domain.service.PlaylistService;
 import com.github.enteraname74.musik.domain.utils.ServiceResult;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/playlist")
 public class PlaylistController {
     private final PlaylistService playlistService;
+    private final AuthService authService;
 
     @Autowired
-    public PlaylistController(PlaylistService playlistService) {
+    public PlaylistController(PlaylistService playlistService, AuthService authService) {
         this.playlistService = playlistService;
+        this.authService = authService;
     }
 
+    /**
+     * Create a playlist.
+     *
+     * @param playlist the playlist to create.
+     * @return a ResponseEntity, with the response of the request.
+     */
     @PostMapping()
     public ResponseEntity<?> add(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @RequestBody Playlist playlist
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
         ServiceResult<?> result =  playlistService.save(playlist);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
@@ -34,18 +46,30 @@ public class PlaylistController {
      * @return a list containing all Playlists or an empty list if there is an error.
      */
     @GetMapping("/all")
-    List<Playlist> getAll() { return playlistService.getAll(); }
+    ResponseEntity<?> getAll(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token
+    ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
+
+        return new ResponseEntity<>(
+                playlistService.getAll(),
+                HttpStatus.OK
+        );
+    }
 
     /**
-     * Retrieves a Music from its id.
+     * Retrieves a Playlist from its id.
      *
-     * @param id the id of the Music to retrieve.
-     * @return a ResponseEntity, with the found Music or an error.
+     * @param id the id of the Playlist to retrieve.
+     * @return a ResponseEntity, with the found Playlist or an error.
      */
     @GetMapping("/{id}")
     ResponseEntity<?> get(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @PathVariable String id
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
+
         ServiceResult<?> result = playlistService.getById(id);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
@@ -60,8 +84,10 @@ public class PlaylistController {
      */
     @DeleteMapping("/{id}")
     ResponseEntity<?> delete(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @PathVariable String id
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
         ServiceResult<?> result = playlistService.deleteById(id);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
@@ -76,9 +102,11 @@ public class PlaylistController {
      */
     @GetMapping("/add/{playlistId}/{musicId}")
     ResponseEntity<?> addMusicToPlaylist(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @PathVariable String playlistId,
             @PathVariable String musicId
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
         ServiceResult<?> result = playlistService.addMusicToPlaylist(playlistId, musicId);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
@@ -93,9 +121,11 @@ public class PlaylistController {
      */
     @GetMapping("/remove/{playlistId}/{musicId}")
     ResponseEntity<?> removeMusicFromPlaylist(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @PathVariable String playlistId,
             @PathVariable String musicId
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
         ServiceResult<?> result = playlistService.removeMusicToPlaylist(playlistId, musicId);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
